Squashed commit for larger proofs and separating the OpenCL program
commit 9956f38a29ee7faeae5d430372cf4e72509cb3fb
Author: Guilherme Lawless <guilherme@nano.org>
Date:   Tue Sep 10 17:57:08 2019 +0100

    Fix opencl_program definition

commit 18390ba28487b5c9ea0b0a67406b1e33a32c9395
Author: Guilherme Lawless <guilherme@nano.org>
Date:   Tue Sep 10 17:44:42 2019 +0100

    Moving source to a subdirectory and opencl_program at compile time

commit 8406cebbf81e56d429d485c03b91f71ef91d26b4
Merge: 4393bf7 ec9bef7
Author: Guilherme Lawless <guilherme@nano.org>
Date:   Tue Sep 10 14:03:20 2019 +0100

    Merge branch 'master' into pow_model

commit 4393bf7dfc4974bda2f6d9c83c9de7ba10fa71d8
Author: clemahieu <clemahieu@gmail.com>
Date:   Mon Sep 9 09:39:27 2019 -0500

    Adding the notion value size to pow header. This allows larger memory slabs to be used.

+++ include/nano_pow/cpp_driver.hpp
@@ -76,7 +76,7 @@ namespace nano_pow
 		uint64_t difficulty_inv;
 		uint32_t fill_count () const;
 		size_t size { 0 };
-		uint32_t * slab { nullptr };
+		uint8_t * slab { nullptr };
 	public:
 		std::array<uint64_t, 2> nonce { { 0, 0 } };
 	};
